refactor: 회원쿠폰 발급에 대한 책임소재를 변경하여 순환참조 제거
